refactor(provenance): move get_claim_fields from core to provenance
Pure relocation. The function answers "which fields are claim-controlled?"
— a provenance-domain question. It lived in apps/core/models.py only as
historical accretion, and that placement blocks tightening its parameter
to type[ClaimControlledModel] (the next step in the model-driven metadata
plan) without violating the AppBoundaries rule that core depends on
nothing.

Moves _CLAIMS_EXEMPT_NAMES and get_claim_fields to a new
apps/provenance/models/introspection.py, re-exported from
apps.provenance.models. No signature or behavior change.
