+++ backend/apps/catalog/tests/test_source_enabled.py
@@ -18,9 +18,8 @@ from apps.catalog.resolve import (
 )
 from apps.catalog.resolve._relationships import resolve_all_credits
 from apps.catalog.tests.conftest import make_machine_model
-from apps.core.models import get_claim_fields
 from apps.provenance.helpers import active_claims
-from apps.provenance.models import Claim, Source
+from apps.provenance.models import Claim, Source, get_claim_fields
 
 User = get_user_model()
